refactor: create reusable confirmation dialog component
- Add ConfirmationDialog component with configurable props for title, message, buttons
- Replace DeleteConfirmation and DeleteQuizConfirmation with reusable component
- Reduce code duplication by ~160 lines across both confirmation dialogs
- Support HTML in messages and flexible onSuccess/invalidateQueries patterns
- Maintain all existing functionality while improving maintainability
