+++ frontend/src/components/UserSettings/DeleteConfirmation.tsx
@@ -1,108 +1,22 @@
-import { Button, ButtonGroup, Text } from "@chakra-ui/react"
-import { useMutation, useQueryClient } from "@tanstack/react-query"
-import { useState } from "react"
-import { useForm } from "react-hook-form"
-
-import { type ApiError, UsersService } from "@/client"
-import {
-  DialogActionTrigger,
-  DialogBody,
-  DialogCloseTrigger,
-  DialogContent,
-  DialogFooter,
-  DialogHeader,
-  DialogRoot,
-  DialogTitle,
-  DialogTrigger,
-} from "@/components/ui/dialog"
-import { useCustomToast, useErrorHandler } from "@/hooks/common"
+import { UsersService } from "@/client"
+import ConfirmationDialog from "@/components/ui/confirmation-dialog"
 import { useAuth } from "@/hooks/auth"
 
 const DeleteConfirmation = () => {
-  const [isOpen, setIsOpen] = useState(false)
-  const queryClient = useQueryClient()
-  const { showSuccessToast } = useCustomToast()
-  const { handleError } = useErrorHandler()
-  const {
-    handleSubmit,
-    formState: { isSubmitting },
-  } = useForm()
   const { logout } = useAuth()
 
-  const mutation = useMutation({
-    mutationFn: () => UsersService.deleteUserMe(),
-    onSuccess: () => {
-      showSuccessToast("Your account has been successfully deleted")
-      setIsOpen(false)
-      logout()
-    },
-    onError: (err: ApiError) => {
-      handleError(err)
-    },
-    onSettled: () => {
-      queryClient.invalidateQueries({ queryKey: ["currentUser"] })
-    },
-  })
-
-  const onSubmit = async () => {
-    mutation.mutate()
-  }
-
   return (
-    <>
-      <DialogRoot
-        size={{ base: "xs", md: "md" }}
-        role="alertdialog"
-        placement="center"
-        open={isOpen}
-        onOpenChange={({ open }) => setIsOpen(open)}
-      >
-        <DialogTrigger asChild>
-          <Button variant="solid" colorPalette="red" mt={4}>
-            Delete
-          </Button>
-        </DialogTrigger>
-
-        <DialogContent>
-          <form onSubmit={handleSubmit(onSubmit)}>
-            <DialogCloseTrigger />
-            <DialogHeader>
-              <DialogTitle>Confirmation Required</DialogTitle>
-            </DialogHeader>
-            <DialogBody>
-              <Text mb={4}>
-                All your account data will be{" "}
-                <strong>permanently deleted.</strong> If you are sure, please
-                click <strong>"Confirm"</strong> to proceed. This action cannot
-                be undone.
-              </Text>
-            </DialogBody>
-
-            <DialogFooter gap={2}>
-              <ButtonGroup>
-                <DialogActionTrigger asChild>
-                  <Button
-                    variant="subtle"
-                    colorPalette="gray"
-                    disabled={isSubmitting}
-                  >
-                    Cancel
-                  </Button>
-                </DialogActionTrigger>
-                <Button
-                  variant="solid"
-                  colorPalette="red"
-                  type="submit"
-                  loading={isSubmitting}
-                >
-                  Delete
-                </Button>
-              </ButtonGroup>
-            </DialogFooter>
-          </form>
-        </DialogContent>
-      </DialogRoot>
-    </>
+    <ConfirmationDialog
+      triggerButtonText="Delete"
+      triggerButtonProps={{ mt: 4 }}
+      title="Confirmation Required"
+      message="All your account data will be <strong>permanently deleted.</strong> If you are sure, please click <strong>'Delete'</strong> to proceed. This action cannot be undone."
+      confirmButtonText="Delete"
+      successMessage="Your account has been successfully deleted"
+      mutationFn={() => UsersService.deleteUserMe()}
+      onSuccess={logout}
+      invalidateQueries={[["currentUser"]]}
+    />
   )
 }
